Rename PDVideoPlayerView to PDVideoPlayer and demote sample
